finish refactor ngestion pipeline
- Add ocr extraction in ingestion pipeline to extract pixel perfect bounding_box as a replacement to llm text extraction

- Modify schema models; add bounding_box and page_number to sync with database

- Add a step to pipeplinr to mark document status as failed when the pipeline throw an error

- Integrate docker

+++ core/api/schema_public_latest.py
@@ -126,7 +126,7 @@ class DocumentLineItemsInsert(CustomModelInsert):
 
     # Optional fields
     gl_code: str | None = Field(default=None)
-    quantity: int | None = Field(default=None)
+    quantity: float | None = Field(default=None)
     unit_price: Decimal | None = Field(default=None)
     page_number: int | None = Field(default=None)
     bounding_box: list[float] | None = Field(default=None)
@@ -238,7 +238,7 @@ class DocumentLineItemsUpdate(CustomModelUpdate):
     description: str | None = Field(default=None)
     document_id: UUID4 | None = Field(default=None)
     gl_code: str | None = Field(default=None)
-    quantity: int | None = Field(default=None)
+    quantity: float | None = Field(default=None)
     total_price: Decimal | None = Field(default=None)
     unit_price: Decimal | None = Field(default=None)
     page_number: int | None = Field(default=None)
